# swaggerJson.json
{
    "swagger": "2.0",
    "info": {
        "title": "compastrips-backend",
        "description": "compastrips Admin API",
        "version": "1.0.0",
        "license": {
            "name": "",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/api/v1",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/user/auth/signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "get all user",
                "operationId": "register",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "first_name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": ""
                        }
                    },
                    {
                        "in": "query",
                        "name": "last_name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": ""
                        }
                    },
                    {
                        "in": "query",
                        "name": "user_name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": ""
                        }
                    },
                    {
                        "in": "query",
                        "name": "email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": ""
                        }
                    },
                    {
                        "in": "query",
                        "name": "password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": ""
                        }
                    },
                    {
                        "in": "query",
                        "name": "gender",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "1 for ascending order AND 0 for descending order"
                        }
                    },
                    {
                        "in": "query",
                        "name": "nationality",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": ""
                        }
                    },
                    {
                        "in": "query",
                        "name": "mobile",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": ""
                        }
                    },
                    {
                        "in": "query",
                        "name": "dob",
                        "required": true,
                        "schema": {
                            "type": "date",
                            "description": ""
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "get all user",
                "operationId": "register",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "email",
                        "required": true,
                        "type": "string",
                        "description": ""
                    },
                    {
                        "in": "query",
                        "name": "password",
                        "required": true,
                        "type": "string",
                        "description": ""
                    }
                ],
                "responses": {
                    "200": {
                        "descrption": "successful operation",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}